Add support for legacy Puppetfile DSL
The Puppetfile DSL supported by R10K supports a few other methods
besides `mod`. Specifically the `forge` method, while potentially being
deprecated, is still in wide use and supported for backwards
compatibility reasons in the R10K DSL. Since it is not supported in The
Forgeinator currently, retrieving modules based on a control repository is
likely to fail for most users.

This commit adds backwards compatibility for the `forge` method. This
method is effectively a noop method and will have no effect on The
Forgeinator's operations. The method is purely to be compatible with the
R10K Puppetfile DSL.