Correct failing spec tests
The spec tests for the Forgeinator::Puppetfile class were written
assuming a global logging facility. This global logging appears to have
been removed previously and replaced with `puts` methods. This change
was not reflected in the spec tests and as such they are failing.

This commit refactors the spec tests to properly look for the output to
STDOUT. The tests are now all passing.